# backend/swagger-ui/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Baustellenverwaltung",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/constructionSite": {
      "get": {
        "tags": [
          "constructionSite"
        ],
        "summary": "Returns all construction sites",
        "description": "Returns a array of all construction sites",
        "operationId": "getConstructionSites",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConstructionSite"
                  },
                  "x-content-type": "application/json"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-swagger-router-controller": "ConstructionSite"
      },
      "post": {
        "tags": [
          "constructionSite"
        ],
        "summary": "Add constructionSite",
        "operationId": "addConstructionSite",
        "requestBody": {
          "description": "ConstructionSite object that should be added",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionSite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionSite"
                }
              }
            }
          },
          "400": {
            "description": "Invalid properties on ConstructionSite",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-codegen-request-body-name": "body",
        "x-swagger-router-controller": "ConstructionSite"
      }
    },
    "/constructionSite/{constructionSiteId}": {
      "get": {
        "tags": [
          "constructionSite"
        ],
        "summary": "Returns construction site",
        "description": "Returns construction site",
        "operationId": "getConstructionSiteById",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSite",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConstructionSite"
                  },
                  "x-content-type": "application/json"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-swagger-router-controller": "ConstructionSite"
      },
      "put": {
        "tags": [
          "constructionSite"
        ],
        "summary": "Update constructionSite",
        "description": "Update Construction site",
        "operationId": "updateConstructionSite",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSite",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "ConstructionSite object that should be added",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionSite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionSite"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "404": {
            "description": "Construction site not found",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-codegen-request-body-name": "body",
        "x-swagger-router-controller": "ConstructionSite"
      },
      "delete": {
        "tags": [
          "constructionSite"
        ],
        "summary": "Delete construction site by ID",
        "description": "Deletes construction Site",
        "operationId": "deleteConstructionSite",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSite",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted",
            "content": {}
          },
          "400": {
            "description": "Invalid parameters",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "404": {
            "description": "Construction site not found",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-swagger-router-controller": "ConstructionSite"
      }
    },
    "/constructionSite/{constructionSiteId}/image": {
      "get": {
        "tags": [
          "constructionSite"
        ],
        "summary": "Returns array with image paths",
        "description": "Returns a array that includes all image paths",
        "operationId": "getConstructionSiteImages",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSiteId",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type":"string"
                  },
                  "x-content-type": "application/json"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-swagger-router-controller": "ConstructionSite"
      },
      "post": {
        "tags": [
          "constructionSite"
        ],
        "summary": "uploads an image",
        "operationId": "uploadImageToConstructionSite",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSiteId",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-swagger-router-controller": "ConstructionSite"
      }
    },
...
    "/constructionSite/{constructionSiteId}/worker": {
      "post": {
        "tags": [
          "constructionSite"
        ],
        "summary": "Add worker to constructionSite",
        "operationId": "addWorker",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSite",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "maximum": 10,
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Id of users that should be added to constructionSite",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
				        "minItems": 1,
				        "items": {
					        "type": "integer",
                  "example": "12, 32, 45"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "minItems": 1,
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "id": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid userId",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
		  "404": {
            "description": "UserId not found",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-codegen-request-body-name": "body",
        "x-swagger-router-controller": "ConstructionSite"
      }
    },
    "/constructionSite/{constructionSiteId}/worker{workerId}": {
      "delete": {
        "tags": [
          "constructionSite"
        ],
        "summary": "Remove worker from Construction site by ID",
        "description": "Removes worker from construction site",
        "operationId": "removeWorker",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSite",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "workerId",
            "in": "path",
            "description": "ID of user that should be removed",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted",
            "content": {}
          },
          "400": {
            "description": "Invalid parameters",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "404": {
            "description": "Worker not found",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-swagger-router-controller": "ConstructionSite"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Contact": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "zip": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "phone": {
            "type": "number"
          }
        },
        "example": {
          "zip": "zip",
          "address": "address",
          "city": "city",
          "phone": 6.027456183070403,
          "name": "name",
          "email": "email"
        },
        "xml": {
          "name": "Contact"
        }
      },
      "ConstructionSite": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdBy": {
            "$ref": "#/components/schemas/Task_createdBy"
          },
          "contact": {
            "$ref": "#/components/schemas/Contact"
          },
          "name": {
            "type": "string",
            "example": "alpina Innenfarbe - weiss"
          },
          "modifiedAt": {
            "type": "string",
            "example": "{}"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "completed"
            ]
          }
        },
        "example": {
          "createdBy": {
            "user": "name",
            "userId": "userId"
          },
          "modifiedAt": "{}",
          "contact": {
            "zip": "zip",
            "address": "address",
            "city": "city",
            "phone": 6.027456183070403,
            "name": "name",
            "email": "email"
          },
          "name": "testbaustelle - 123",
          "id": 0,
          "status": "open",
          "users": [{
            "userId": 1,
            "workeremail": "test123@testmail.de",
            "workername": "yusuf abi"
          }]
        },
        "xml": {
          "name": "ConstructionSite"
        }
      },
...
      "body": {
        "properties": {
          "additionalMetadata": {
            "type": "string",
            "description": "Additional data to pass to server"
          },
          "file": {
            "type": "string",
            "description": "file to upload",
            "format": "binary"
          }
        }
      },
      "Task_createdBy": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "description": "name of user"
          },
          "userId": {
            "type": "string",
            "description": "userId reference"
          }
        },
        "example": {
          "name": "name",
          "userId": "userId"
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}